feat: Adds a shell script to generate a collage.
I took the images that were on `frontend/src/pages/Home.js` and used them for the
collage. I think that it is less complex and more efficient to display one image
on the page instead of multiples.

I envision this as being a background process that runs every 'x' number hours
and gets 'y' number of images from our database, API call, or wherever and
creates the collage.

Additionally, the `frontend/src/components/CardsHolder.js` and
`frontend/src/css/Cards.css` files have been changed to display the
collage.